feat: move kanban components to ui/kanban components
